+++ src/components/diagram/graphCore.tsx
@@ -2,269 +2,25 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import cloneDeep from 'lodash-es/cloneDeep';
 import { Graph, Cell, Markup, Node, Interp, Model, Registry, ObjectExt } from '@antv/x6';
-import { ShareRegistry } from '@antv/x6/lib/model/registry';
+import { FixedClipboard } from './Clipboard';
 import { Options as GraphOptions } from '@antv/x6/lib/graph/options';
-import { ReactShape } from '@antv/x6-react-shape';
+import { ExtNode } from './Node';
 import { EdgeView, NodeView } from '@antv/x6';
 
 import { stencils } from './stencils/index';
 import { StencilEditor } from './stencils/StencilEditor';
 import { EditableCellTool } from './stencils/NodeField';
 import { validateEmbedding, validateConnection } from './interactionValidation';
+import { FixedHistory } from './History';
 
-export class ExtNode extends ReactShape {
-  translate(tx = 0, ty = 0, options: Node.TranslateOptions = {}) {
-    const selectedNodes = this.model?.graph?.getSelectedCells();
-
-    if (this.checkTranslationOwner(options.translateBy) && options.parentCall !== (this.getParent() as Node).id) {
-      return this;
-    }
-    if (tx === 0 && ty === 0) {
-      return this;
-    }
-    if (this.store.get().movable === false && !options.parentCall) {
-      if (this._parent) {
-        this._parent.translate(tx, ty, options);
-      }
-      return this;
-    }
-
-    options.translateBy = options.translateBy || this.id;
-
-    const position = this.getPosition();
-
-    if (options.restrict != null && options.translateBy === this.id) {
-      const bbox = this.getBBox({ deep: true });
-      const ra = options.restrict;
-      const dx = position.x - bbox.x;
-      const dy = position.y - bbox.y;
-      const x = Math.max(ra.x + dx, Math.min(ra.x + ra.width + dx - bbox.width, position.x + tx));
-      const y = Math.max(ra.y + dy, Math.min(ra.y + ra.height + dy - bbox.height, position.y + ty));
-
-      tx = x - position.x;
-      ty = y - position.y;
-    }
-
-    const translatedPosition = {
-      x: position.x + tx,
-      y: position.y + ty,
-    };
-
-    options.tx = tx;
-    options.ty = ty;
-
-    if (options.transition) {
-      if (typeof options.transition !== 'object') {
-        options.transition = {};
-      }
-
-      this.transition('position', translatedPosition, {
-        ...options.transition,
-        interp: Interp.object,
-      });
-      const newOptions = { ...options, parentCall: true };
-      this.eachChild((child) => {
-        if (selectedNodes?.indexOf(child) === -1) {
-          child.translate(tx, ty, newOptions);
-        }
-      });
-    } else {
-      this.startBatch('translate', options);
-      this.store.set('position', translatedPosition, options);
-      options.handledTranslation = options.handledTranslation || [];
-      if (!options.handledTranslation.includes(this)) {
-        options.handledTranslation.push(this);
-        const children = this._children;
-        if (children?.length) {
-          options.handledTranslation.push(...children);
-        }
-      }
-      const newOptions = { ...options, parentCall: this.id };
-      this.eachChild((child) => {
-        child.translate(tx, ty, newOptions);
-      });
-      this.stopBatch('translate', options);
-    }
-
-    return this;
-  }
-  getStore() {
-    return this.store;
-  }
-  checkTranslationOwner(ownerId: string | number | undefined): boolean {
-    let child = this as Node;
-    while (child.getParent()) {
-      if ((child.getParent() as Node).id === ownerId) {
-        return true;
-      }
-      child = child.getParent() as Node;
-    }
-    return false;
-  }
-  eachChild(iterator: (child: Cell, index: number, children: Cell[]) => void, context?: any) {
-    (this._children || []).forEach(iterator, context);
-    return this;
-  }
-  getChildCount() {
-    return this._children?.length || 0;
-  }
-  insertChild(child: Cell | null, index?: number, options: Cell.SetOptions = {}): this {
-    if (child != null && child !== this) {
-      const oldParent = child.getParent();
-      const changed = this !== oldParent;
-
-      let pos = index;
-      if (pos == null) {
-        pos = this.getChildCount();
-        if (!changed) {
-          pos -= 1;
-        }
-      }
-
-      // remove from old parent
-      if (oldParent) {
-        const children = oldParent.getChildren();
-        if (children) {
-          const index = children.indexOf(child);
-          if (index >= 0) {
-            child.setParent(null, options);
-            children.splice(index, 1);
-            oldParent.setChildren(children, options);
-          }
-        }
-      }
-      let children = this._children;
-      if (children == null) {
-        children = [];
-        children.push(child);
-      } else {
-        children.splice(pos, 0, child);
-      }
-
-      child.setParent(this, options);
-      this.setChildren(children, options);
-
-      if (changed && this.model) {
-        const incomings = this.model.getIncomingEdges(this);
-        const outgoings = this.model.getOutgoingEdges(this);
-
-        if (incomings) {
-          incomings.forEach((edge) => edge.updateParent(options));
-        }
-
-        if (outgoings) {
-          outgoings.forEach((edge) => edge.updateParent(options));
-        }
-      }
-
-      if (this.model) {
-        this.model.addCell(child, options);
-      }
-    }
-
-    return this;
-  }
-  toJSON(options: Cell.ToJSONOptions = {}): Node.Properties {
-    const props = { ...this.store.get() };
-    const toString = Object.prototype.toString;
-    const cellType = 'node';
-
-    if (!props.shape) {
-      const ctor = this.constructor;
-      throw new Error(
-        `Unable to serialize ${cellType} missing "shape" prop, check the ${cellType} "${
-          ctor.name || toString.call(ctor)
-        }"`,
-      );
-    }
-
-    const ctor = this.constructor as typeof Cell;
-    const diff = options.diff === true;
-    const attrs = props.attrs || {};
-    const presets = ctor.getDefaults(true) as Node.Properties;
-    // When `options.diff` is `true`, we should process the custom options,
-    // such as `width`, `height` etc. to ensure the comparing work correctly.
-    const defaults = diff ? this.preprocess(presets, true) : presets;
-    const defaultAttrs = defaults.attrs || {};
-    const finalAttrs: any = {};
-
-    Object.keys(props).forEach((key) => {
-      if (key !== 'yogaProps') {
-        const val = props[key];
-        if (val != null && !Array.isArray(val) && typeof val === 'object' && !ObjectExt.isPlainObject(val)) {
-          throw new Error(
-            `Can only serialize ${cellType} with plain-object props, but got a "${toString.call(
-              val,
-            )}" type of key "${key}" on ${cellType} "${this.id}"`,
-          );
-        }
-
-        if (key !== 'attrs' && key !== 'shape' && diff) {
-          const preset = defaults[key];
-          if (ObjectExt.isEqual(val, preset)) {
-            delete props[key];
-          }
-        }
-      } else {
-        delete props[key];
-      }
-    });
-
-    Object.keys(attrs).forEach((key) => {
-      const attr = attrs[key];
-      const defaultAttr = defaultAttrs[key];
-
-      Object.keys(attr).forEach((name) => {
-        const value = attr[name] as any;
-        const defaultValue = defaultAttr ? defaultAttr[name] : null;
-
-        if (value != null && typeof value === 'object' && !Array.isArray(value)) {
-          Object.keys(value).forEach((subName) => {
-            const subValue = value[subName];
-            if (
-              defaultAttr == null ||
-              defaultValue == null ||
-              !ObjectExt.isObject(defaultValue) ||
-              !ObjectExt.isEqual(defaultValue[subName], subValue)
-            ) {
-              if (finalAttrs[key] == null) {
-                finalAttrs[key] = {};
-              }
-              if (finalAttrs[key][name] == null) {
-                finalAttrs[key][name] = {};
-              }
-              const tmp = finalAttrs[key][name] as any;
-              tmp[subName] = subValue;
-            }
-          });
-        } else if (defaultAttr == null || !ObjectExt.isEqual(defaultValue, value)) {
-          // `value` is not an object, default attribute with `key` does not
-          // exist or it is different than the attribute value set on the cell.
-          if (finalAttrs[key] == null) {
-            finalAttrs[key] = {};
-          }
-          finalAttrs[key][name] = value as any;
-        }
-      });
-    });
-
-    const finalProps = {
-      ...props,
-      attrs: ObjectExt.isEmpty(finalAttrs) ? undefined : finalAttrs,
-    };
-
-    if (finalProps.attrs == null) {
-      delete finalProps.attrs;
-    }
-
-    const ret = finalProps as any;
-    if (ret.angle === 0) {
-      delete ret.angle;
-    }
-
-    return ObjectExt.cloneDeep(ret);
+export class ExtGraph extends Graph {
+  public readonly history: FixedHistory;
+  constructor(options: Partial<GraphOptions.Manual>) {
+    super(options);
+    this.history = new FixedHistory({ graph: this, ...this.options.history });
   }
 }
+
 class SimpleNodeView extends NodeView {
   protected renderMarkup() {
     this.renderJSONMarkup({
@@ -379,7 +135,7 @@ export const createGraph = ({
     // typically happens during recompilation
     console.log(e);
   }
-  const g = new Graph({
+  const g = new ExtGraph({
     container: refContainer.current,
     width: width,
     height: height,
@@ -389,7 +145,11 @@ export const createGraph = ({
     resizing: {
       enabled: true,
     },
-    history: true,
+    history: {
+      executeCommand: () => {
+        console.log('command');
+      },
+    },
     clipboard: {
       enabled: true,
     },
@@ -492,7 +252,9 @@ export const createGraph = ({
       arrowheadMovable: true,
     },
   });
-
+  const clipboard = g.clipboard;
+  clipboard.widget = new FixedClipboard();
+  clipboard.widget.deserialize(clipboard.instanceOptions);
   g.on('selection:changed', ({ added, removed }: { added: Cell[]; removed: Cell[] }) => {
     added.forEach((cell) => {
       if (cell.isEdge()) {
@@ -675,7 +437,7 @@ export const addNewEdges = ({ graph, edgesData }) => {
   });
 };
 
-export const nodeFromData = ({ data, shape, Renderer }) => ({
+export const nodeFromData = ({ data, shape, Renderer, onSave = () => null, setEditing = () => null }) => ({
   id: data['@id'],
   size: { width: data.width, height: data.height },
   position: { x: data.x, y: data.y },
@@ -683,15 +445,18 @@ export const nodeFromData = ({ data, shape, Renderer }) => ({
   style: data.style,
   shape: shape,
   attrs: data.attrs,
+  label: data.label || data?.subject?.title || `${data?.subject?.name}: ${data?.subject?.datatype}`,
   zIndex: data.zIndex,
   movable: data.movable !== undefined ? data.movable : true,
   editing: false,
   component(n) {
     const setEditing = (state: boolean) => {
+      n._model.graph.history.disable();
       n.setProp('editing', state);
+      n._model.graph.history.enable();
     };
     const onSave = (t: string) => {
-      n.setProp('editing', false);
+      setEditing(false);
       n.setProp('label', t);
     };
     return <Renderer node={n} data={cloneDeep(n.store.data)} setEditing={setEditing} nodeData={data} onSave={onSave} />;
