+++ src/components/diagram/stencils/NodeField.tsx
@@ -24,6 +24,7 @@ const fieldStyle: React.CSSProperties = {
   textOverflow: 'ellipsis',
   border: 0,
   boxShadow: 'none !important',
+  backgroundColor: 'white',
 };
 
 export const NodeField = React.memo(
@@ -57,7 +58,7 @@ export const NodeField = React.memo(
     );
   },
   (prev: any, next: any) => {
-    if (prev.data?.editing !== next.data?.editing) {
+    if (prev.data?.editing !== next.data?.editing || prev.data?.label !== next.data?.label) {
       return false;
     }
     return true;
